// src/generated/subscription/v1/subscription.ts
export interface SubscribeRequest {
  email: string;
  repo: string;
}

export interface SubscribeResponse {}

export interface ConfirmSubscriptionRequest {
  token: string;
}

export interface ConfirmSubscriptionResponse {}

export interface UnsubscribeRequest {
  token: string;
}

export interface UnsubscribeResponse {}

export interface GetSubscriptionsRequest {
  email: string;
}

export interface SubscriptionItem {
  email: string;
  repo: string;
  confirmed: boolean;
  lastSeenTag: string;
}

export interface GetSubscriptionsResponse {
  subscriptions: SubscriptionItem[];
}

function createBaseSubscribeRequest(): SubscribeRequest {
  return { email: '', repo: '' };
}

export const SubscribeRequest: MessageFns<SubscribeRequest> = {
  encode(
    message: SubscribeRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.email !== '') {
      writer.uint32(10).string(message.email);
    }
    if (message.repo !== '') {
      writer.uint32(18).string(message.repo);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SubscribeRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSubscribeRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.email = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.repo = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<SubscribeRequest>, I>>(
    base?: I,
  ): SubscribeRequest {
    return SubscribeRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SubscribeRequest>, I>>(
    object: I,
  ): SubscribeRequest {
    const message = createBaseSubscribeRequest();
    message.email = object.email ?? '';
    message.repo = object.repo ?? '';
    return message;
  },
};

function createBaseSubscribeResponse(): SubscribeResponse {
  return {};
}

export const SubscribeResponse: MessageFns<SubscribeResponse> = {
  encode(
    _: SubscribeResponse,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SubscribeResponse {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSubscribeResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<SubscribeResponse>, I>>(
    base?: I,
  ): SubscribeResponse {
    return SubscribeResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SubscribeResponse>, I>>(
    _: I,
  ): SubscribeResponse {
    const message = createBaseSubscribeResponse();
    return message;
  },
};

function createBaseConfirmSubscriptionRequest(): ConfirmSubscriptionRequest {
  return { token: '' };
}

export const ConfirmSubscriptionRequest: MessageFns<ConfirmSubscriptionRequest> =
  {
    encode(
      message: ConfirmSubscriptionRequest,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.token !== '') {
        writer.uint32(10).string(message.token);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): ConfirmSubscriptionRequest {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseConfirmSubscriptionRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.token = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    create<I extends Exact<DeepPartial<ConfirmSubscriptionRequest>, I>>(
      base?: I,
    ): ConfirmSubscriptionRequest {
      return ConfirmSubscriptionRequest.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<ConfirmSubscriptionRequest>, I>>(
      object: I,
    ): ConfirmSubscriptionRequest {
      const message = createBaseConfirmSubscriptionRequest();
      message.token = object.token ?? '';
      return message;
    },
  };

function createBaseConfirmSubscriptionResponse(): ConfirmSubscriptionResponse {
  return {};
}

export const ConfirmSubscriptionResponse: MessageFns<ConfirmSubscriptionResponse> =
  {
    encode(
      _: ConfirmSubscriptionResponse,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): ConfirmSubscriptionResponse {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseConfirmSubscriptionResponse();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    create<I extends Exact<DeepPartial<ConfirmSubscriptionResponse>, I>>(
      base?: I,
    ): ConfirmSubscriptionResponse {
      return ConfirmSubscriptionResponse.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<ConfirmSubscriptionResponse>, I>>(
      _: I,
    ): ConfirmSubscriptionResponse {
      const message = createBaseConfirmSubscriptionResponse();
      return message;
    },
  };

function createBaseUnsubscribeRequest(): UnsubscribeRequest {
  return { token: '' };
}

export const UnsubscribeRequest: MessageFns<UnsubscribeRequest> = {
  encode(
    message: UnsubscribeRequest,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.token !== '') {
      writer.uint32(10).string(message.token);
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): UnsubscribeRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUnsubscribeRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.token = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<UnsubscribeRequest>, I>>(
    base?: I,
  ): UnsubscribeRequest {
    return UnsubscribeRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UnsubscribeRequest>, I>>(
    object: I,
  ): UnsubscribeRequest {
    const message = createBaseUnsubscribeRequest();
    message.token = object.token ?? '';
    return message;
  },
};

function createBaseUnsubscribeResponse(): UnsubscribeResponse {
  return {};
}

export const UnsubscribeResponse: MessageFns<UnsubscribeResponse> = {
  encode(
    _: UnsubscribeResponse,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): UnsubscribeResponse {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseUnsubscribeResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<UnsubscribeResponse>, I>>(
    base?: I,
  ): UnsubscribeResponse {
    return UnsubscribeResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<UnsubscribeResponse>, I>>(
    _: I,
  ): UnsubscribeResponse {
    const message = createBaseUnsubscribeResponse();
    return message;
  },
};
// ...
export type SubscriptionServiceService = typeof SubscriptionServiceService;
export const SubscriptionServiceService = {
  subscribe: {
    path: '/subscription.v1.SubscriptionService/Subscribe' as const,
    requestStream: false as const,
    responseStream: false as const,
    requestSerialize: (value: SubscribeRequest): Buffer =>
      Buffer.from(SubscribeRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): SubscribeRequest =>
      SubscribeRequest.decode(value),
    responseSerialize: (value: SubscribeResponse): Buffer =>
      Buffer.from(SubscribeResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): SubscribeResponse =>
      SubscribeResponse.decode(value),
  },
  confirmSubscription: {
    path: '/subscription.v1.SubscriptionService/ConfirmSubscription' as const,
    requestStream: false as const,
    responseStream: false as const,
    requestSerialize: (value: ConfirmSubscriptionRequest): Buffer =>
      Buffer.from(ConfirmSubscriptionRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): ConfirmSubscriptionRequest =>
      ConfirmSubscriptionRequest.decode(value),
    responseSerialize: (value: ConfirmSubscriptionResponse): Buffer =>
      Buffer.from(ConfirmSubscriptionResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): ConfirmSubscriptionResponse =>
      ConfirmSubscriptionResponse.decode(value),
  },
  unsubscribe: {
    path: '/subscription.v1.SubscriptionService/Unsubscribe' as const,
    requestStream: false as const,
    responseStream: false as const,
    requestSerialize: (value: UnsubscribeRequest): Buffer =>
      Buffer.from(UnsubscribeRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): UnsubscribeRequest =>
      UnsubscribeRequest.decode(value),
    responseSerialize: (value: UnsubscribeResponse): Buffer =>
      Buffer.from(UnsubscribeResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): UnsubscribeResponse =>
      UnsubscribeResponse.decode(value),
  },
  getSubscriptions: {
    path: '/subscription.v1.SubscriptionService/GetSubscriptions' as const,
    requestStream: false as const,
    responseStream: false as const,
    requestSerialize: (value: GetSubscriptionsRequest): Buffer =>
      Buffer.from(GetSubscriptionsRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer): GetSubscriptionsRequest =>
      GetSubscriptionsRequest.decode(value),
    responseSerialize: (value: GetSubscriptionsResponse): Buffer =>
      Buffer.from(GetSubscriptionsResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer): GetSubscriptionsResponse =>
      GetSubscriptionsResponse.decode(value),
  },
} as const;
// ...
export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}